+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 - Updated `notan_egui` to the latest version of `egui` 0.19.
 - Added mipmapping support with `TextureBuilder::generate_mipmaps`.
+- Added `WindowBackend::position` and `WindowBackend::set_position`.
 
 ## v0.7.1 - 08/10/2022